Bug #23523869 IMPLEMENT NDB_PERROR AS REPLACEMENT FOR "PERROR --NDB"
Overview:

"perror" is a server tool which takes error code as input and prints the
error message. It prints ndb error codes when --ndb is passed as
argument.

Problem:

As perror is a server tool, when the mysql is not compiled with cluster
flags then we will not get the ndb error codes. So, vanilla MySQL Server
when compiled will not have "perror --ndb" functionality.

Solution:

Implemented new tool ndb_perror with functionality identical to that of
perror.

Signed-off-by: Dinesh Surya Prakash <dinesh.prakash@oracle.com>