Bug#24658707 ASSERT: BUF0BUF.CC:2469:BUF_BLOCK_GET_STATE(BLOCK) == BUF_BLOCK_FILE_PAGE
Issue
=====
The issue is that during commit_inplace_alter_table we wake up the purge
thread and we take a btr search latch and try to disable the adaptive hash
search system and empty the index. The purge operation happening in the
background, when flushing the pages, tries to remove possible adaptive hash
index on the page and it sets the the block state as BUF_BLOCK_REMOVE_HASH
and waits on the btr search latch taken by the alter command. And in the
alter command when we're trying to empty the hash index of the same block
we hit the assertion ut_ad(buf_block_get_state(block) ==
BUF_BLOCK_FILE_PAGE) as the block state was changed by the purge operation.
Both the threads would be working on the same block at the time of
assertion.

Fix
===
The fix has already been pushed to trunk as part of meta-sync branch.
Readding the assert which was removed earlier to silence the pb2 failures
which was because of this bug.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>