WL#9536: Add a 'version' field to innodb_dynamic_metadata, also dict_table_t::version accordingly.
This version is used to check if the dynamic metadata is of the correct
version to be applied, to support crash-safe dynamic metadata persistence.

Code clean-up so not to write back to DDTableBuffer on DDLs.

Also after fix the bug of reading autoinc from dd::Table::se_private_data,
remove the remembering of autoinc during open() functions.