Bug#16371292 : Print FK violation information (PART I : ndbd + ndbapi fix)
When there is an error in a SQL query on an NDB table due to a Foreign
Key constraint violation, the details of the foreign key are not shown
in the error message. The server actually asks the ndb handler about
the FK details before printing it but the handler itself doesn't know
the FK details, so, it sends back a 'Unknown error code' message. In
fact the violation information never leaves the DBTC block.

This patch fixes the ndbd and ndbapi side to handle and return the fk
id on a fk constraint violation.

The DBDICT block now sets the apiConnectPtr's errorData variable to the
fk id on failure. Rest of the code is already in place. This fk id
reaches back to the ndbapi via the TCROLLBACKREP signal and gets read.
The ndbapi adds this additional details to the transactions's ndberror.

Also now during a drop table failure due to an FK constraint, the fk id
is saved into the NdbDictionary's ndberror by the ndbapi.

This patch also updates the getNdbErrorDetails function of the ndb so
that now it returns the Foreign Key's fully qualified name when an
NdbError filled with a FK constraint violation error is sent to it.