wl#10234 Replace temp FragLocation array with pool allocated FragLocationRec.
The previous wl#10234 patch set introduced ScanRecord::fragLocations[256]
to manage the set of fragments to be scanned. As the size of the
ScanRecord used to be ~100bytes, adding this array of FragLocations
increased it size by a factor of ~20. This was never intended
to be a permanent part of this wl, so this patch change the
fragment locations to be allocated from a dynamic pool of
fragment location records.

The records are allocated as we get the location info from DHDIH
and are released as soon as the fragment scan request has been
sent to LQH.

The fragment location records are added to the tail of a SLFifoList
and consumed from its head in order to operate on them in the
same order as they are received from DBDIH.