wl#10234 Removed the now obsolete member variable ScanFragRec::scanFragId.
Previously used as an argument sent to DIH in GSN_<insert signal name>.
However usage of this signal was replaced with GSN_DIGETNODESREQ
which does not take such a 'fragId' argument.

The few remaining uses of this variable was partly incorrect,
(Fixed by prev patches in this wl) or could easily be replaced
by checking scanNextFragId instead.