Bug #25997748: MIGRATE FROM HASH TO STD::UNORDERED_MAP [patch 13, noclose]
More conversions from HASH: The remaining client code, and SAFE_HASH,
used in MyISAM's key cache. The key cache should of course also go away
(this simply converts it to a string -> string map, which is essentially
just as untyped as using HASH), but MyISAM is really the only user,
so we're letting the core logic stay.

Change-Id: I794f80a208871d4f43caeb6d4fc2547b99bfa784