Bug#26239591 ALLOW JOIN-PUSHDOWN OF JOIN IN COMBINATION WITH 'GROUP BY'
Due to previous unclear seperation between the optimize
and execute phases when the query involved a 'GROUP BY',
the 'join-pushable evaluator' couldn't previously be sure
whether the optimized QEP (Query Execution Plan) was pushable
or not. Thus it choosed to not push 'grouped joins'.

This has been cleaned up by WL5558 long ago, so we can
now remove this limitation.