Bug #25997748: MIGRATE FROM HASH TO STD::UNORDERED_MAP [patch 15, noclose]
Remove the last instances of HASH (except the query cache, which will be
fixed in a pending worklog). In particular, this fixes a bug where
partition names would not be correctly compared, as the old hash code assumed
that two strings with a different number of bytes could not compare equal.

Change-Id: Ibb8ff1650490ed6db5443342da789a098e33b210