WL#9536: To fix Bug#26277872 - ASSERT IN MODE == 16 || !FIX_BLOCK->PAGE.FILE_PAGE_WAS_FREED AT BUF_PAGE_GET_GEN
In ALTER TABLE, we may drop the newly created indexes immediately during
rollback. However, the postDDL will also try to free this kind of indexes.
Thus it may lead to double free issue as this bug says.

The solution is not to drop the new index in the ALTER TABLE rollback
phase, just do it in postDDL phase.