Bug#2603593 GROUP_REPLICATION_GROUP_NAME SHOULD NOT MATCH SERVER_UUID
Description:
When a node requests to join the group, we don't check to see if the joiner
node's @@server_uuid == @@group_replication_group_name.

That can lead to a variety of serious issues because we rely on GTIDs being
globally unique and the basic idea that each unique GTID (UUID:GNO) refers to
the same underlying data changes. But in the above scenario UUID:N could have
been generated for one user transaction on the stand-alone instance. And it
could have been generated for an entirely different user transaction in the
group. That's because the UUID part of the GTID that both independent systems
were generating was the same.