Bug#25068254:: GENERATE ERRORS IF SPECIAL STRINGS ARE PART OF THE VAR DIRECTORY PATH
Description:
============
There is a function in mysqltest.cc called fix_win_paths() which replaces back
slashes with front slashes in paths when --replace_result is called in the test.
The replacement of slashes happens when a certain pattern is matched in the path

The list of patterns is defined in mysqltest.cc under init_win_path_patterns()

There is a problem when any of these patterns is present in the vardir path
unintentionally because all the back slashes get replaced when it is not
supposed to. It was observed when running tests like main.secure_file_priv_win and
i_main.bug11761752 on Windows.

perl mysql-test-run.pl main.secure_file_priv_win --vardir=undo [fail]

Solution:
=========
Changes in the string patterns are made which are used in order to find paths

Reviewed by:
Pavan Naik <pavan.naik@oracle.com>
RB: 16254