wl#10234, fix fragment enumeration skew. (performance patch)
To avoid that all SPJ blocks requested fragment scans
on the same sub-set of fragments, while other were
idle, a different 'skew' is used wjen setting up the
fragment lists on the different SPJ's.

This was previously set up such that a skew of '1'
was used between the different SPJ.

With these WL that has to change such that the number
of multi-fragments assigned to each SPJ is used in
the skew calculation.