wl#10234: Deprecate usage of 'm_firstExecution' member variable
... compliments previous patch.

'm_firstExecution' was used as a flag to handle that we didn't
have valid 'parallelism' statistics yet. However, it usage
was a bit obfuscated due to its naming.

This patch introduce the ::isValid() method in the class
handling the statistics. Usage of 'm_firstExecution' can then
be replaced with calling ::isValid() instead.