BUG#26314472: SERVER STARTUP CRASH IF IT STARTED WITH GROUP_REPLICATION COMMANDS IN INIT-FILE
Group Replication does communicate with the server through the SQL
API, to which depending on the thread context from which the
communication starts, the plugin may require to launch a thread.

When that thread was launched, a memory pointer was released but was
not protected against double free, this patch fixes that.