Bug#25738624: ASSERTION `FALSE' FAILED IN SQL/SQL_EXECUTOR.CC
An assertion protecting an unused branch failed if an indexed column in
a system table was referenced in a MIN/MAX and in a join, and the
optimizer was unable to optimize away the join in opt_sum_query(). If an
index was used to extract the MIN/MAX value of the column, the row
buffer of that table would get status 'started' with the content stored
in the record[0] buffer. If the optimizer later tried to access the
table using read_system(), this status would cause an attempt to restore
the content from the record[1] buffer instead of reading it directly
from the table. Because only record[0] contains valid data and
the branch restoring the data from record[1] is declared as unused, the
assertion fails as it should.

To account for cases where opt_sum_query() is able to determine MIN/MAX
value from an indexed column in a system table but is not able to
optimize away subsequent reads to the same table, the row buffer status
is now reset to 'not started' after reading the index. Any subsequent
read will then fetch data from the table instead of attempting to
restore it from record[1].