BUG#26148011 RPL.RPL_MULTI_SOURCE_CORRUPT_REPOSITORY              FAILS WITH RESULT CONTENT MISMATCH
DESCRIPTION
===========
The test uses show_relaylog_events.inc to show the content of
some relay logs. the relay logs sometimes have different
content. That caused test failures.

ANALYSIS
========
The reason is that FLUSH LOGS is called immediately after
slave is started. At the time, slave may or may not
get Rotate/Format_description events from master. Most of
the time, the events arrive before FLUSH LOGS and are written
into the relay logs which are not showed by the test. But
they sometime arrive just after FLUSH LOGS, so they are
written to the relay logs which are showed by the test.

FIX
===
The test just cares if SHOW RELAYLOG EVENTS FOR CHANNEL can
be executed successfully, but not the result of the statement.
So we just ignore the result by using --disable_result_log.