Bug#26349496 [ERROR] INNODB: ASSERTION FAILURE: MACH0DATA.IC:[0-9]+:B
Issue
=====
The issue is that if there is an incomplete redo log, then during parsing
of the dict log record (mlog_parse_initial_dict_log_record()) it's possible
that the log record ptr is made null during the fetching of the table id
info. This is not checked before we proceed to version fetch which results
in this bug.

Fix
===
Addded a ptr==nullptr check after id fetch in
mlog_parse_initial_dict_log_record().

Reviewed-by: Bin Su <bin.x.su@oracle.com>