WL#8592 SRID type modifier for geometric types
This worklog adds a new column property for geometric types to specify the
SRID:

  CREATE TABLE t1 (g GEOMETRY SRID 4326, p POINT SRID 0 NOT NULL);

Values inserted into a column with an SRID property must be in that SRID.
Attempts to insert values with other SRIDs results in an exception condition
being raised. Unmodified types, i.e., types with no SRID specification, will
continue to accept all SRIDs, as before.

Change-Id: I144c65c4d6cd4d434ebd4a6062b4d34980c7f1bd