Bug#26117735: MYSQLBINLOG READ-FROM-REMOTE-SERVER NOT HONORING REWRITE_DB FILTERING
Problem:
=======
When running mysqlbinlog with the --read-from-remote-server
option, the processing of the binlog entries accepts
rewrite_db as an argument, but ignores the rewrite rule.
Therefore no filtering is done on the binlog entries,
resulting in failures to write data in the target DB.

Analysis:
========
mysqlbinlog tool can read binary logs that are present
locally or on a remote machine. At present rewrite filter is
applied for binary logs that are read locally. This
rewrite_db filter specific code is not invoked while dumping
binary log from a remote machine. Hence the filtering is is
not done on those binlog entries.

Fix:
===
Invoke rewrite_db filter code while dumping remote log
entries as well.