BUG#26021345: MAKE MTR PICK UP COLLECTIONS/DISABLED-<SANITIZER>.LIST FILES
Description:
------------
Tests to be skipped temporarily in a sanitizer are listed in
'mysql-test/collections/disabled-<sanitizer>.list'. MTR should pick
up these files automatically when running with --sanitize.

When it does so, it should output the list of tests being skipped
and the name of the disabled-file. If run with --enable-disabled, it
shouldn't pick up the file.

Fix:
----
When '--sanitize' option is enabled, MTR automatically picks up
mysql-test/collections/disabled-<sanitizer>.list file and adds it to
the list of files containing tests to be skipped in the current run.

Reviewed-by: Deepa Dixit <deepa.dixit@oracle.com>
Reviewed-by: Srikanth B R <srikanth.b.r@oracle.com>
RB: 16172