Bug#26162562 DISABLING PERFORMANCE_SCHEMA.SETUP_OBJECT DROP TABLE STATS
Before this fix,

when a table instrumentation is disabled,
due to edits in table performance_schema.setup_objects,
the following actions were executed:
- the table lock statistics were discarded
- the index statistics were discarded
- the index names, stored with the per index stats, were lost.

The root cause is logic present in
PFS_table_share::refresh_setup_object_flags(),
which explicitly destroyed lock and index stats.

Note that the table statistics were implemented with the following model:
- either a table is enabled, and stats are preserved,
- or a table is disabled, and stats are dropped.

This model however was not consistently implemented,
as statistics could still be allocated even for disabled table,
for example in find_or_create_table_share().

With this fix, the old model is abandoned,
as the behavior is inconsistent with other instrumentations,
and cause undesirable side effects for index names.

The new model is as follows:
- enabling / disabling instrumentation for tables
  no longer destroys associated statistics.

As a result,
- index names are preserved once found,
  so that index statistics no longer show spurious "index N" names.
- statistics are preserved instead of being truncated
  when a table is disabled.
- visiting table statistics searches data deeper,
  no longer stopping if a table instrumentation is disabled.