Bug#26357806 SESSION_CONNECT_ATTRS.PROCESSLIST_ID SHOULD BE BIGINT UNSIGNED
Patch for 8.0 only.

Before this fix, columns:
- performance_schema.session_connect_attrs.PROCESSLIST_ID
- performance_schema.session_account_connect_attrs.PROCESSLIST_ID

are of type INT.

This is incorrect, as column threads.PROCESSLIST_ID
is a BIGINT UNSIGNED.

The root cause is the incomplete change introduced by
  Bug#14664453 THREAD_ID SHOULD BE BIGINT(21) UNSIGNED IN PERFORMANCE_SCHEMA
to migrate thread and processlist ids from 32 bit to 64 bits.

With this fix, columns
- performance_schema.session_connect_attrs.PROCESSLIST_ID
- performance_schema.session_account_connect_attrs.PROCESSLIST_ID
are changed to the proper type.