Bug#25380000: SOME WARNINGS APPEAR IN DUMP FROM MYSQLDUMP
Description:- Deprecation warning messages are printed
along with the dump data in the result file for mysqldump
utility.

Analysis:- Deprecation warning messages are printed to
STDOUT instead of STDERR for all client utilities which
uses SSL options to connect to server.

Fix:- Warning messages are printed to STDERR instead of
STDOUT so that these messages will not appear in the
result file populated by redirecting STDOUT.