WL#11298 - MySQL GCS: Support name resolution in whitelist in MySQL 5.7
EXECUTIVE SUMMARY
======================
This worklog implements support for hostnames in the
GCS/XCom connection whitelisting feature. Therefore,
users will be able to configure whitelisting not only
with IPs but also with hostmames

DESCRIPTION
======================

In Cloud Environments, the following scenarios are quite common:
- Machines popping on and off of a group;
- IP Address are obtained from a pool that is dedicated to some machines;

In this case, it is quite hard to maintain a proper whitelist, since values are in
constant mutation.

For that, it was requested that, since we already support names in peer addresses
and in local addresses, one should also support names in the whitelist parameters,
e.g: www.randomname.com/16, as an example.

User Story:
- As a DBA, i want to be able to configure whitelist parameter using hostnames
instead of IP Addresses so that I am able to allow connections from outside
hosts that do not have a fixed address

- As a DBA, i want to be able to configure whitelist parameter using a mix of
hostnames and IP Addresses with netmasks I am able to allow connections from
outside
hosts that do not have a fixed address and still allow fixed IPs to connect to
the group