Bug#26327996 WL#9536 CRASH DURING ALTER ENCRYPTION='N' OR '' TABLE NOT ACCESSIBLE
Crash happend after ibd files is renamed in row_rename_table_for_mysql,
and checkpoint, before dd info in server layer is committed.
In this scenario, boot_tablespaces will use the old space id to load the
tablespace, and fail. So, in ddl replay stage, it can't find the
fil_space_t obj to do rename ibd file operation.
Solution: we use the info in tablespace.open.* to open the ibd file and
load the fil_space_t obj to the cache.