Bug#26405350: MAIN.MYSQLHOTCOPY_ARCHIVE AND               MAIN.MYSQLHOTCOPY_MYISAM TESTS               FAILING IN PB2-5.6
Description: Mtr tests, main.mysqlhotcopy_archive and
main.mysqlhotcopy_myisam are failing recently in PB2-5.6.

Analysis: The test failure is due to the mismatch in the
exit code returned by the mysqlhotcopy utility when the
utility encounters an error.  This mismatch is happening due
to the unpredictable exit code returned by the "die" command
used in the "mysqlhotcopy" perl script.

In the perl documentation for "die" command it is mentioned
that "the value of the exit code used by die can be
non-predictable, so should not be relied upon, other than
to be non-zero."

Fix: Error number 255 is also added along with 9, 11, 110
and 2304.