Bug #24938374   MYSQL CRASHED AFTER LONG WAIT ON DICT OPERATION LOCK WHILE SYNCING FTS INDEX
PROBLEM

As part of rb#12340, dict_operation_lock was acquired to avoid dropping
of FTS index/ table while the sync is in progress in the background. Due
to this change server gets killed on long wait for dict_operation_lock
while a long sync is in progress.

SOLUTION

We remove the dict operation lock that is held by sync (fts) in
background. We will manage the concurrency of drop index and sync index
in 2 ways:
1) If sync is already in progress, Drop index would wait for sync to
complete.
2) If alter is already happening and sync is invoked then sync would do
a check for index/table to be dropped flags are set or not, if it is it
would skip syncing that index.

Reviewed by: Jimmy Yang <Jimmy.Yang@oracle.com>
RB: 15344