WL#10793 - InnoDB: Use VATS for scheduling lock release under high load
Variance-Aware Transaction Scheduling (VATS) is described in
https://arxiv.org/pdf/1602.01871.pdf. The patch was contributed by
Jiamin Huang <jiamin@umich.edu> and Barzan Mozafari <mozafari@umich.edu>
from the University of Michigan.

When granting locks the grant strategy used depends on the number of waiting
locks. If it is >= 32 then VATS will be used otherwise the original first come
first served (FCFS) strategy will be used. The threshold of 32 was chosen
after empirical verification on a supra class machine. Ideally we would
like to default to VATS for all loads but a small regression was seen on
the lower loads.

Reviewed-by: Jimmy <jimmy.yang@oracle.com>
Reviewed-by: Deb <debarun.banerjee@oracle.com>
RB:16290