Bug#25407964 GROUP BY DESC GIVES WRONG RESULT WHEN GROUPS ON DECIMAL AND SEES A NULL
Cause: when grouping we have to compare the values of group expressions
to see when we enter a new group. Here the expression is of DECIMAL
type, and the comparison of old-value non-NULL with new-value NULL
(as happens with DESC, where NULLs are last), was not working:
cached_item_decimal::cmp() returned FALSE, meaning "new and old are equal",
so two groups were merged into one.

Fix: make cmp() return TRUE. This is similar to cached_item_int::cmp().