BUG#26093967: AVOID SETTING GROUP_REPLICATION_FORCE_MEMBERS IN A PARTITION THAT HOLDS MAJORITY
group_replication_force_members option was introduced has a
break-the-glass measure to unblock groups that did lost
communication with a majority of members. This safety measure allows
a DBA or control tool to quickly resume operations on a unblocked
group by forcing a subset of the previous group membership.

Unfortunately, this option was allowed to be used on situations on
which the group was working properly, that is, a majority was
reachable. This was never the purpose of this option, and such
action could cause instability on the group.
Therefore, we are restricting its use to the scenario to which it
was created: for a subset of previous membership when a majority of
the members are unreachable.

This is an optimist solution, as the majority of members can change
right after the is_majority_unreachable returns.