Bug#25432850 - MYSQLPUMP: PROGRESS INFORMATION INCORRECT                FOR TABLES
DESCRIPTION
===========
While mysqlpump is performing backup, it shows incorrect
progress of the tables.

For example for a database having 500 tables each having
500 rows, mysqlpump progress looks as shown below:

Dump progress: 0/1 tables, 250/500 rows
Dump progress: 0/73 tables, 18250/36500 rows
Dump progress: 0/151 tables, 37750/75500 rows
Dump progress: 0/229 tables, 57250/114500 rows
Dump progress: 0/307 tables, 76750/153500 rows
Dump progress: 0/385 tables, 96250/192500 rows
Dump progress: 0/463 tables, 115750/231500 rows
Dump completed in 9241 milliseconds

Please note that the progress information of the
total-no-of-tables-backed-up at any given point of time
remains zero throughout the process.

ANALYSIS
========
To report progress of both Tables and Rows,
Abstract_progress_watcher::object_processing_ended()
calls
Abstract_progress_watcher::progress_changed()

When doing it for Tables, the 1st function checks if the
'processed_table_task' is completed. If not, it currently
skips calling 2nd function and as a result the progress
report for the Table isn't shown.

On removing the check, we see the progress appearing for
both Tables and Rows. Moreover the check is not required as
the other conditions are sufficient in themselves to pass
the criteria.

FIX
===
Drop the above condition to check if the
'processed_table_task' is completed.