Bug#26375914: SOME TESTS WITH MAJORITY LOSS FAIL SPORADICALLY IN PB2
Some tests where failing as they do not suppress the messages that
members output when they are unreachable.
We chose to use a global suppression though, as more failures were
also seen on Valgrind runs due to resource starvation caused
partitions.
We use this bug even so to do proper cleanup on one of these tests
that was missing it.