Bug #26269280: SET SESSION BINLOG_FORMAT = ROW GIVES SYNTAX ERROR
Problem: in MySQL 8, ROW had been added to the list of reserved words
but ROW is also one of the values that SET command should be able to
set to variables like "binlog_format".
Since the "SET binlog_format = ROW" case is very common, but ROW
is a reserved word now, it has been decided to introduce an exception
in the syntax to accept ROW as before as a value in SET assignments.