Bug#26483625: DON'T PRODUCE LOGICAL DIFFS FOR NO-OP UPDATES
When partial update of JSON values is used, and JSON_SET changes a
value in a JSON document, and the new value is exactly the same as the
old value, a logical diff may be collected.

There is code that avoids collection of binary diffs for these no-ops.
This patch extends that code so that it also avoids the unnecessary
collection of logical diffs.

The patch also removes some leftover debug code from a unit test.