Bug#26495028: REMOVE UNNECESSARY JSON CODE FROM BINARY
Use function objects instead of function pointers as arguments to
std::find_if_not and std::any_of in json_path.cc. Even though GCC
seems to be able to inline the function calls in release builds with
optimization level 3, it leaves the function definitions in the
binary, presumably because pointers to them have been taken. Using
function objects makes it eliminate the function definitions from the
binary. An additional benefit is that function objects are more likely
to be inlined at lower optimization levels.

Also change the calculation of the typelit_max_length constant in
item_json_func.cc to use constexpr functions so that it gets
calculated during compilation.

Change-Id: Id90f64a9a68f894c688905a75df5adb74f7e3971